Update profilephoto-configure-settings.md
It is not clear in the document which roles are supported for usage with the PhotoUpdateSettings API, however it is clear that only a small set of built-in roles will work for this. In my testing, seems to be Global Administrator, User Administrator, and People Administrator. We should add a note for clarity.